+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Cloudflare Turnstile support for bot protection
+- Delete account functionality
+- Forgot password functionality
+- Change password functionality
 
 ### Changed
 - By minumum require 8 characters for password